+++ src/main/java/org/monarchinitiative/lr2pg/cmd/SimulateSvgPhenoOnlyCommand.java
@@ -10,8 +10,8 @@ import org.monarchinitiative.phenol.formats.hpo.HpoOntology;
 import org.monarchinitiative.phenol.ontology.data.TermId;
 
 /**
- * Simulate a single case and produce an SVG output file.
- * TODO -- make output file name flexible.
+ * Simulate a single case and produce an SVG org.monarchinitiative.lr2pg.output file.
+ * TODO -- make org.monarchinitiative.lr2pg.output file name flexible.
  * @author <a href="mailto:peter.robinson@jax.org">Peter Robinson</a>
  */
 public class SimulateSvgPhenoOnlyCommand extends SimulatePhenotypesCommand {
@@ -20,14 +20,14 @@ public class SimulateSvgPhenoOnlyCommand extends SimulatePhenotypesCommand {
 
     public SimulateSvgPhenoOnlyCommand(String dataDirPath, String diseaseId){
         super(dataDirPath);
-        this.diseaseCurie=TermId.constructWithPrefix(diseaseId);
+        this.diseaseCurie=TermId.of(diseaseId);
     }
 
     public void run() throws Lr2pgException {
         PhenotypeOnlyHpoCaseSimulator phenotypeOnlyHpoCaseSimulator = getPhenotypeOnlySimulator();
         HpoDisease disease = phenotypeOnlyHpoCaseSimulator.name2disease(diseaseCurie);
         if (disease==null) {
-            throw new Lr2pgException("Could not find disease for " + diseaseCurie.getIdWithPrefix());
+            throw new Lr2pgException("Could not find disease for " + diseaseCurie.getValue());
         }
         phenotypeOnlyHpoCaseSimulator.simulateCase(disease);
         HpoCase hpocase = phenotypeOnlyHpoCaseSimulator.getCurrentCase();