+++ src/main/java/org/monarchinitiative/lr2pg/gt2git/package-info.java
@@ -1,4 +1,7 @@
 /**
- * Classes related to calculating genotype frequency to genic intolerance.
+ * Classes related to calculating the distribution of pathogenicity scores of ClinVar or all GNOMAD variants.
+ * These classes are not needed to run the main LR2PG algorithm, but instead org.monarchinitiative.lr2pg.output data as a file that is used
+ * to enable the genotype likelihood ratio calculations in the main LR2PG algorithm.
+ * @author <a href="mailto:peter.robinson@jax.org">Peter Robinson</a>
  */
 package org.monarchinitiative.lr2pg.gt2git;