+++ src/test/java/org/monarchinitiative/lr2pg/hpo/HpoCaseTest.java
@@ -6,8 +6,9 @@ import com.google.common.collect.ImmutableMap;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import org.junit.BeforeClass;
-import org.junit.Test;
+
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
 import org.monarchinitiative.lr2pg.likelihoodratio.PhenotypeLikelihoodRatio;
 import org.monarchinitiative.lr2pg.likelihoodratio.PhenotypeLikelihoodRatioTest;
 import org.monarchinitiative.lr2pg.likelihoodratio.TestResult;
@@ -17,20 +18,20 @@ import org.monarchinitiative.phenol.formats.hpo.HpoOntology;
 import org.monarchinitiative.phenol.io.obo.hpo.HpOboParser;
 import org.monarchinitiative.phenol.io.obo.hpo.HpoDiseaseAnnotationParser;
 import org.monarchinitiative.phenol.ontology.data.TermId;
-import org.monarchinitiative.phenol.ontology.data.TermPrefix;
+
 
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.Map;
 
-import static junit.framework.TestCase.assertNotNull;
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 
 /**
  * Test whether we can successfully create HpoCaseOld objects.
  */
-public class HpoCaseTest {
+class HpoCaseTest {
     private static final Logger logger = LogManager.getLogger();
     /** Name of the disease we are simulating in this test, i.e., OMIM:108500. */
     private static String diseasename="108500";
@@ -39,8 +40,8 @@ public class HpoCaseTest {
     private static PhenotypeLikelihoodRatio backforeFreq;
 
 
-    @BeforeClass
-    public static void setup() throws PhenolException, FileNotFoundException,NullPointerException {
+    @BeforeAll
+    static void setup() throws PhenolException, FileNotFoundException,NullPointerException {
         ClassLoader classLoader = PhenotypeLikelihoodRatioTest.class.getClassLoader();
         String hpoPath = classLoader.getResource("hp.obo").getFile();
         String annotationPath = classLoader.getResource("small.hpoa").getFile();
@@ -52,12 +53,12 @@ public class HpoCaseTest {
         backforeFreq=new PhenotypeLikelihoodRatio(ontology,diseaseMap);
 
         /* these are the phenotypic abnormalties of our "case" */
-        TermPrefix HP_PREFIX=new TermPrefix("HP");
-        TermId t1 = new TermId(HP_PREFIX,"0006855");
-        TermId t2 = new TermId(HP_PREFIX,"0000651");
-        TermId t3 = new TermId(HP_PREFIX,"0010545");
-        TermId t4 = new TermId(HP_PREFIX,"0001260");
-        TermId t5 = new TermId(HP_PREFIX,"0001332");
+        String HP_PREFIX="HP";
+        TermId t1 = TermId.of(HP_PREFIX,"0006855");
+        TermId t2 = TermId.of(HP_PREFIX,"0000651");
+        TermId t3 = TermId.of(HP_PREFIX,"0010545");
+        TermId t4 = TermId.of(HP_PREFIX,"0001260");
+        TermId t5 = TermId.of(HP_PREFIX,"0001332");
         ImmutableList.Builder<TermId> builder = new ImmutableList.Builder<>();
         builder.add(t1,t2,t3,t4,t5);
         // We need to provide a list of TestResult objects for the API, but they are not required for this unit test
@@ -71,7 +72,7 @@ public class HpoCaseTest {
 
 
     @Test
-    public void testNotNull() {
+    void testNotNull() {
         assertNotNull(hpocase);
     }
 
@@ -85,7 +86,7 @@ public class HpoCaseTest {
      Thus there are five Hpo annotations.
      */
     @Test
-    public void testNumberOfAnnotations() {
+    void testNumberOfAnnotations() {
         int expected=5;
         assertEquals(expected,hpocase.getNumberOfObservations());
     }