+++ src/test/java/org/monarchinitiative/lr2pg/likelihoodratio/TestResultTest.java
@@ -1,9 +1,9 @@
 package org.monarchinitiative.lr2pg.likelihoodratio;
 
 import com.google.common.collect.ImmutableList;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.monarchinitiative.phenol.formats.hpo.HpoAnnotation;
 import org.monarchinitiative.phenol.formats.hpo.HpoDisease;
 import org.monarchinitiative.phenol.ontology.data.TermId;
@@ -12,14 +12,15 @@ import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 
 /**
  * Some of this test class is based on the data and cases presented in
  * https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2683447/
  * Note -- the authors of that paper rounded results and this class does not!
  */
-public class TestResultTest {
+class TestResultTest {
 
     private static final double EPSILON = 0.00001;
     private static double ratio(double sensitivity, double specificity) {
@@ -27,16 +28,16 @@ public class TestResultTest {
     }
     private HpoDisease glaucoma;
 
-    @Before
-    public void init() {
-        TermId glaucomaId = TermId.constructWithPrefix("MONDO:123");
+    @BeforeEach
+    void init() {
+        TermId glaucomaId = TermId.of("MONDO:123");
         List<TermId> emptyList = ImmutableList.of();
         List<HpoAnnotation> emptyAnnot = ImmutableList.of();
         glaucoma = new HpoDisease("Glaucoma",glaucomaId,emptyAnnot,emptyList,emptyList);
     }
 
     @Test
-    public void testGlaucomaLR1() {
+    void testGlaucomaLR1() {
         TestResult tresult;
 
         ImmutableList.Builder<Double> builder = new ImmutableList.Builder<>();
@@ -66,9 +67,9 @@ public class TestResultTest {
     }
 
     @Test
-    public void testGlaucomaLR2() {
+    void testGlaucomaLR2() {
         TestResult tresult;
-        TermId glaucomaId = TermId.constructWithPrefix("MONDO:123");
+        TermId glaucomaId = TermId.of("MONDO:123");
         ImmutableList.Builder<Double> builder = new ImmutableList.Builder<>();
         // The prevalence of glaucoma is 2.5%
         double prevalence = 0.025;
@@ -92,9 +93,9 @@ public class TestResultTest {
 
 
     @Test
-    public void testCompositepositiveLR() {
+    void testCompositepositiveLR() {
         TestResult tresult;
-        TermId glaucomaId = TermId.constructWithPrefix("MONDO:123");
+        TermId glaucomaId = TermId.of("MONDO:123");
         ImmutableList.Builder<Double> builder = new ImmutableList.Builder<>();
         // The prevalence of glaucoma is 2.5%
         double prevalence = 0.025;
@@ -111,25 +112,25 @@ public class TestResultTest {
         tresult = new TestResult(builder.build(),glaucoma,prevalence);
          //PretestOdds = pretest prob / (1-pretest prob) = 0.95 / 0.05 = 19.0
         double expectedPretestOdds = 0.0256410;
-        Assert.assertEquals(expectedPretestOdds, tresult.pretestodds(), EPSILON);
+        assertEquals(expectedPretestOdds, tresult.pretestodds(), EPSILON);
 
         //PosttestOdds = PrestestOdds * Compositelikelihoodratio = 0.03 * 6.25 * 20 * 20
         double expected = 64.102564;
-        Assert.assertEquals(expected, tresult.posttestodds(), EPSILON);
+        assertEquals(expected, tresult.posttestodds(), EPSILON);
 
         //PosttestProb = PosttestOdds / (1+ PosttestOdds)
         double ptodds=expected;
         expected = 0.9846396;
-        Assert.assertEquals(expected, (ptodds/(ptodds+1)), EPSILON);
+        assertEquals(expected, (ptodds/(ptodds+1)), EPSILON);
     }
 
 
     @Test
-    public void testTestResultSorting() {
+    void testTestResultSorting() {
         double EPSILON=0.0001;
-        TermId testId1 = TermId.constructWithPrefix("MONDO:1");
-        TermId testId2 = TermId.constructWithPrefix("MONDO:2");
-        TermId testId3 = TermId.constructWithPrefix("MONDO:3");
+        TermId testId1 = TermId.of("MONDO:1");
+        TermId testId2 = TermId.of("MONDO:2");
+        TermId testId3 = TermId.of("MONDO:3");
         List<TermId> emptyList = ImmutableList.of();
         List<HpoAnnotation> emptyAnnot = ImmutableList.of();
         HpoDisease d1 = new HpoDisease("d1",testId1,emptyAnnot,emptyList,emptyList);
@@ -157,7 +158,7 @@ public class TestResultTest {
         // now add another test result, same as result3 but with additional genotype evidence
         // result4 should now be the top hit
         double genotypeLR=2.0;
-        TermId geneId=TermId.constructWithPrefix("NCBI:Faks");
+        TermId geneId=TermId.of("NCBI:Fake");
         TestResult result4=new TestResult(list3,d3,genotypeLR,geneId,prevalence);
         lst.add(result4);
         assertEquals(lst.get(3),result4);